requirements.txt
decorator==4.0.9
defusedxml==0.4.1
Django==1.9.4
django-allauth==0.24.1
django-cors-headers==1.1.0
django-registration==2.0.4
gunicorn==19.4.5
ipython==4.1.2
ipython-genutils==0.1.0
oauthlib==1.0.3
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.8
ptyprocess==0.5.1
python3-openid==3.0.9
pytz==2016.1
PyYAML==3.13
requests==2.14.2
requests-oauthlib==0.6.1
simplegeneric==0.8.1
traitlets==4.1.0
factory_boy==2.9.0
dependency-injector==4.41.0
fs==2.4.16
docker==4.4.4
PyGithub==1.56
urllib3==1.26.13
freezegun==1.2.2
autopep8==2.0.1
